+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support for Pluggable Database added
 - Support for APM Domain service added
+- Support for Database Migration service added
 
 ## 4.31.0 (June 16, 2021)